Second video of mern stack. Created jwt authentication

+++ MERN/backend/controllers/goalController.js
@@ -1,12 +1,13 @@
 const asyncHandler = require('express-async-handler');
 
 const Goal = require('../models/goalModel');
+const User = require('../models/userModel');
 
 // @desc    Get goals
 // @route   GET /api/goals
 // @access  Private
 const getGoals = asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id });
     res.status(200).json(goals);
 });
 
@@ -16,13 +17,18 @@ const getGoals = asyncHandler(async (req, res) => {
 const setGoal = asyncHandler(async (req, res, next) => {
     const requestBody = req.body || {};
     
+    // Add debugging
+    console.log('req.user:', req.user);
+    console.log('req.body:', req.body);
+
     if (!requestBody.text) {
         res.status(400);
-        return next(new Error('Please add a text field'));
+        throw new Error('Please add a text field');
     }
 
     const goal = await Goal.create({
         text: requestBody.text,
+        user: req.user.id,
     });
 
     res.status(200).json(goal);
@@ -37,6 +43,19 @@ const updateGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400);
         throw new Error('Goal not found');
+    }
+
+    // Check for user
+    const user = await User.findById(req.user.id);
+    if (!user) {
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
     }
 
     const updatedGoal = await Goal.findByIdAndUpdate(req.params.id, req.body, {
@@ -55,6 +74,19 @@ const deleteGoal = asyncHandler(async (req, res) => {
     if (!goal) {
         res.status(400);
         throw new Error('Goal not found');
+    }
+
+    // Check for user
+    const user = await User.findById(req.user.id);
+    if (!user) {
+        res.status(401);
+        throw new Error('User not found');
+    }
+
+    // Make sure the logged in user matches the goal user
+    if (goal.user.toString() !== user.id) {
+        res.status(401);
+        throw new Error('User not authorized');
     }
 
     await goal.deleteOne();
